+++ src/ui/diff.rs
@@ -1,7 +1,7 @@
 use crate::app::{App, SelectedPane};
 
 use ratatui::layout::Constraint;
-use ratatui::style::{Color, Modifier, Style};
+use ratatui::style::{Modifier, Style};
 use ratatui::text;
 use ratatui::widgets::{Block, Borders, Cell, Row, Table};
 
@@ -9,7 +9,7 @@ pub fn render(app: &App) -> Table {
   let container = Block::default()
     .title(title(app))
     .borders(Borders::ALL)
-    .style(Style::default().fg(Color::LightCyan));
+    .style(Style::default().fg(app.theme.border_color()));
 
   let mut rows: Vec<Row> = vec![];
 
@@ -22,10 +22,10 @@ pub fn render(app: &App) -> Table {
         let stale = app.since >= log.commit_datetime;
 
         let cells = [
-          Cell::from(sha(&log.sha)),
+          Cell::from(sha(&log.sha, app)),
           Cell::from(age(&log.commit_datetime, app)),
-          Cell::from(author(&log.author)),
-          Cell::from(message(&log.message)),
+          Cell::from(author(&log.author, app)),
+          Cell::from(message(&log.message, app)),
         ];
 
         Row::new(cells).style(stale_style(stale))
@@ -41,7 +41,7 @@ pub fn render(app: &App) -> Table {
 
   Table::new(rows, widths)
     .block(container)
-    .row_highlight_style(Style::default().add_modifier(Modifier::BOLD))
+    .row_highlight_style(app.theme.diff_row_hightlight_style())
     .highlight_symbol("·")
     .column_spacing(2)
 }
@@ -54,25 +54,25 @@ fn title(app: &App) -> text::Span {
 }
 
 fn age<'a>(datetime: &'a chrono::DateTime<chrono::Utc>, app: &App) -> text::Span<'a> {
-  let text_style = Style::default().fg(Color::Red);
+  let text_style = Style::default().fg(app.theme.diff_age_color());
   let text = datetime
     .with_timezone(&app.timezone_offset)
     .format("%a %b %d %R");
   text::Span::styled(text.to_string(), text_style)
 }
 
-fn author(text: &String) -> text::Span {
-  let text_style = Style::default().fg(Color::Yellow);
+fn author<'a>(text: &'a String, app: &'a App) -> text::Span<'a> {
+  let text_style = Style::default().fg(app.theme.diff_author_color());
   text::Span::styled(text, text_style)
 }
 
-fn message(text: &String) -> text::Span {
-  let text_style = Style::default().fg(Color::White);
+fn message<'a>(text: &'a String, app: &'a App) -> text::Span<'a> {
+  let text_style = Style::default().fg(app.theme.diff_message_color());
   text::Span::styled(text, text_style)
 }
 
-fn sha(text: &String) -> text::Span {
-  let text_style = Style::default();
+fn sha<'a>(text: &'a String, app: &'a App) -> text::Span<'a> {
+  let text_style = Style::default().fg(app.theme.diff_sha_color());
   text::Span::styled(text, text_style)
 }