+++ src/ui/modal.rs
@@ -1,7 +1,7 @@
 use crate::app::{App, SelectedModal};
 use crate::widget::calendar::Calendar;
 use ratatui::layout::{Margin, Rect};
-use ratatui::style::{Color, Style};
+use ratatui::style::Style;
 use ratatui::widgets::{Block, Borders, Clear};
 use ratatui::Frame;
 
@@ -22,7 +22,7 @@ pub fn render(app: &App, frame: &mut Frame) {
       let container = Block::default()
         .title(" Calendar ")
         .borders(Borders::ALL)
-        .style(Style::default().fg(Color::LightCyan));
+        .style(Style::default().fg(app.theme.border_color()));
 
       let calendar = Calendar::new().block(container);